# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Auth Practice API",
    "version": "1.0.0",
    "description": "Authentication and user management API using Supabase Auth and Bearer JWT tokens."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your Supabase JWT access token."
      }
    }
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "summary": "Sign up a new user",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "example": "test@example.com" },
                  "password": { "type": "string", "example": "password123" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created successfully" },
          "400": { "description": "Validation error or registration failed" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Log in and obtain access & refresh tokens",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "example": "test@example.com" },
                  "password": { "type": "string", "example": "password123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful, returns JWT tokens" },
          "400": { "description": "Email or password missing" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Log out the current user",
        "tags": ["Authentication"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": { "description": "User logged out successfully" },
          "401": { "description": "Unauthorized - Missing or invalid token" }
        }
      }
    },
    "/public/info": {
      "get": {
        "summary": "Public information endpoint",
        "tags": ["Public"],
        "responses": {
          "200": { "description": "Returns open public information" }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "summary": "Get authenticated user profile",
        "tags": ["Protected"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": { "description": "Profile data retrieved successfully" },
          "401": { "description": "Unauthorized - Token missing or invalid" }
        }
      }
    },
    "/admin/profile": {
      "get": {
        "summary": "Get admin profile data",
        "description": "Restricted endpoint accessible only to authenticated users with the 'admin' role.",
        "tags": ["Admin"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Admin profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "example": "6d32c567-4144-40db-9323-54e5238f33e1" },
                    "email": { "type": "string", "example": "admin@test.com" },
                    "role": { "type": "string", "example": "admin" },
                    "created_at": { "type": "string", "example": "2026-08-18T20:00:00.000Z" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing, malformed, or expired token"
          },
          "403": {
            "description": "Forbidden - Authenticated user lacks admin role"
          }
        }
      }
    }
  }
}